fix: refactor and improve logic for usage under PHP 8
As it turns out, in PHP 8 when comparing a number and a non-numeric string, the engine casts the number to a string and then compares the strings.
This means that `'b' == 0` or `'1a' == 1` will now fail under PHP 8.
As such, I've updated some assertions to vary based on whether we are testing under PHP 8 and above, or PHP 7.
This also afforded me the opportunity to refactor the `InArray` logic to be more readable, as well as more correct.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>