qa: use officially documented regex (with tweaks)
This patch updates the REGEX_BIC constant to follow the documentation from the bundesbank.
The main changes are that the two characters following the country code have more restrictions (the first cannot contain the digit 1, the second cannot contain the letter "o").
A comment with a link to the documentation (and reference to the page) is provided.

CS was also indicating we had two unused import statements, so I removed those.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>