Add installation requirements for use of the default handler
The default handler, which is used in the example, depends on
laminas-uri, however it is not installed automatically. Adding a brief
note about this and instructions on how to install will help users who
are otherwise confronted with a 'class not found' exception when running
the example.

Installation requirements block copied from csrf.md and modified to
match the URI validator.

Signed-off-by: Paul Waring <paul@phpdeveloper.org.uk>