Add new Validator: `HostWithPublicIPv4Address`
Whilst doing some refactoring for v3, I found that the Email Address validator validates IP addresses/hostnames for reserved IPs.

In order to test this effectively, it makes more sense to implement this as a validator and compose that into the Email address validator - it is likely to prove useful to some users too in a standalone context.

So, this validator is a pretty straightforward extraction of https://github.com/laminas/laminas-validator/blob/2.57.x/src/EmailAddress.php#L282-L344

The expectation is that the functionality in the Email validator will be replaced with this.

Signed-off-by: George Steel <george@net-glue.co.uk>