Firm up Callback arguments and document `$context` parameter.
Changes the Callback validator so that the callable always receives `(mixed $value, array $context, ...$otherArgs)` - this stabilizes the args for all environments, i.e. standalone usage and when part of input-filter etc.

Because the Callback validator is used internally, by the CreditCard validator, this also affects the signature of the `service` option to include context and declares the `$context` param in `CreditCard::isValid()` so that it can correctly be passed to the composed callback validator.

Signed-off-by: George Steel <george@net-glue.co.uk>