Conditionally `get()` validators when options are empty.
Because `get()` no longer supports options (to be compliant with psr-container), validator chain used `build()` to retrieve all validator instances.

This is problematic when the user has configured a validator instance under `services`, because `build()` of-course does not fetch from `services`.

The fix here is to call `get()` when no options are given.

Signed-off-by: George Steel <george@net-glue.co.uk>